+++ packages/ct-router/CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [2.4.1] - 2021-03-19
+### Added
+ - Update @decorators
+ - Add private props
+ - add static styles with css``
+
+## [2.4.0] - 2021-03-16
+### Added
+ - Add `beforeunload` listener
+
 ## [2.2.0] - 2020-10-12
 ### Added
  - Add Typescript types